Resolution when packer absent from syspath
When you are running the command check for packer version, this will obviously
fail when packer hasnt been installed into syspath yet. There is an
ignore_errors to handle this but the next extract task relies on register
variable and attr stdout which doesn't exist. Hence that will fail on ansible
2.2+.